feat: Add a new Pagination component
This component is a layer on top of the PaginationData component to make desktop friendly paginations easier to implement. The benefit over the bgr-pagination component is that the user has full control over the item contents. This makes it possible to add the needed classes / content / components that a projects need, instead of being hard locked into the what bgr-pagination provides. The downside is that it requires more code.